+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-- When using the `x-alpha-request-factory` parameter in combination with a service method which used an attrs or pydantic data class in its signature, the request factory did not correctly process the request. The API model was not properly converted to the expected data class instance. This has been fixed to ensure proper handling of attrs and pydantic data classes when using the request factory. Additional integration tests have been added to verify the correct behavior of the request factory an response factory in combination with attrs and pydantic data classes.
+- When using the `x-alpha-request-factory` parameter in combination with a service method which used an attrs or pydantic data class in its signature, the request factory did not correctly process the request. The API model was not properly converted to the expected data class instance. This has been fixed to ensure proper handling of attrs and pydantic data classes when using the request factory. Additional integration tests have been added to verify the correct behavior of the request factory and response factory in combination with attrs and pydantic data classes.
 
 ## [0.7.7] - 2026-09-10